+++ test/Integration/Gateways/GpApiConnector/DataResidencyTest.php
@@ -2,15 +2,13 @@
 
 namespace Gateways\GpApiConnector;
 
-use GlobalPayments\Api\Entities\Enums\Channel;
-use GlobalPayments\Api\Entities\Enums\DataResidency;
-use GlobalPayments\Api\Entities\Enums\Environment;
-use GlobalPayments\Api\Entities\Enums\ServiceEndpoints;
+use GlobalPayments\Api\Entities\Enums\{Channel, DataResidency, Environment, ServiceEndpoints};
+use GlobalPayments\Api\Entities\Exceptions\GatewayException;
 use GlobalPayments\Api\PaymentMethods\CreditCardData;
-use GlobalPayments\Api\ServiceConfigs\Gateways\GpApiConfig;
-use GlobalPayments\Api\ServiceConfigs\Gateways\PorticoConfig;
+use GlobalPayments\Api\ServiceConfigs\Gateways\{GpApiConfig, PorticoConfig};
 use GlobalPayments\Api\ServicesContainer;
 use GlobalPayments\Api\Tests\Data\BaseGpApiTestConfig;
+use GlobalPayments\Api\Utils\Logging\RequestConsoleLogger;
 use PHPUnit\Framework\TestCase;
 
 class DataResidencyTest extends TestCase
@@ -40,6 +38,7 @@ class DataResidencyTest extends TestCase
         $config->environment = Environment::TEST;
         $config->channel = Channel::CardNotPresent;
         $config->dataResidency = DataResidency::EU;
+        $config->requestLogger = new RequestConsoleLogger();
 
         ServicesContainer::configureService($config, 'eu-config');
         
@@ -62,6 +61,7 @@ class DataResidencyTest extends TestCase
         $config->appKey = BaseGpApiTestConfig::$appKey;
         $config->environment = Environment::TEST;
         $config->channel = Channel::CardNotPresent;
+        $config->requestLogger = new RequestConsoleLogger();
 
         ServicesContainer::configureService($config, 'default-config');
         
@@ -93,9 +93,35 @@ class DataResidencyTest extends TestCase
         $config->environment = GpApiConfig::QA_ENVIRONMENT;
         $config->channel = Channel::CardNotPresent;
         $config->dataResidency = DataResidency::EU;
+        $config->requestLogger = new RequestConsoleLogger();
 
         ServicesContainer::configureService($config, 'qa-eu-config');
         $this->assertEquals(ServiceEndpoints::GP_API_QA_EU, $config->serviceUrl);
+
+        $this->expectException(GatewayException::class);
+        $this->card->charge(1.00)
+            ->withCurrency('USD')
+            ->execute('qa-eu-config');
+    }
+
+    /** Test DataResidency EU + Production environment routes to Production EU endpoint */
+    public function testDataResidencyEuProductionRoutesToProdEuEndpoint(): void
+    {
+        $config = new GpApiConfig();
+        $config->appId = BaseGpApiTestConfig::EU_APP_ID;
+        $config->appKey = BaseGpApiTestConfig::EU_APP_KEY;
+        $config->environment = Environment::PRODUCTION;
+        $config->channel = Channel::CardNotPresent;
+        $config->dataResidency = DataResidency::EU;
+        $config->requestLogger = new RequestConsoleLogger();
+
+        ServicesContainer::configureService($config, 'prod-eu-config');
+        $this->assertEquals(ServiceEndpoints::GP_API_PRODUCTION_EU, $config->serviceUrl);
+
+        $this->expectException(GatewayException::class);
+        $this->card->charge(1.00)
+            ->withCurrency('USD')
+            ->execute('prod-eu-config');
     }
 
     /** Test manually provided serviceUrl overrides residency/environment routing */